requirements.txt
openai
opencv-python
pillow
pdf2image
requests
google-generativeai
fastapi
uvicorn
python-multipart
google-genai
pydantic
pymupdf
opencv-python-headless
numpy
python-dotenv







